Use icons in the buttons for adding/remove telemetry rows/columns
It wasn't intuitive which buttons were for adding rows and which where for adding columns. It would be very annoying if a user accidentally deleted a row when they meant to delete a column. This makes it more the users fault if a mistake is make (very obvious which button does what if you look closesly)

Contribution Sponsor: Firestorm (launchfirestorm.com)