Add "Request Gimbal Control" popup
If another user was in control of the gimbal, the current logic would stop gimbal commands from being sent out, but not show a "Request Gimbal Control" popup. There was a signal being emitted suggesting  a popup was intended to be shown, but wasn't. Now a popup is indeed shown.